Added new animated gif of gameplay, moved old one aside, ignore AVI files.
Added `images/*.avi' to .gitignore, so large video files used in creating the
animated gifs will not be included in the git repo. Moved images/pong.gif to
images/pong-old.gif. Created new images/pong.gif with a smaller size (400x400)
so as not to take up too much space in the README on github.